Decouple api
