+++ src/api/package.json
@@ -1,0 +1,104 @@
+{
+  "name": "mid",
+  "version": "1.0.0",
+  "description": "",
+  "private": true,
+  "scripts": {
+    "build": "rm -rf dist && tsc",
+    "start": "nodemon ./server.ts",
+    "test": "jest",
+    "posttest": "npm run lint",
+    "lint": "eslint \"**/*.{ts,tsx}\"",
+    "lint:fix": "eslint \"**/*.{ts,tsx}\" --fix",
+    "pretest": "npm run setup:db",
+    "prestart": "npm run setup:db",
+    "setup:db": "sqlite3 ../../sql/dev.db < ../../sql/createSchema.sql && sqlite3 ../../sql/dev.db < ../../sql/createData.sql"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "UNLICENSED",
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node",
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js"
+    ]
+  },
+  "eslintConfig": {
+    "root": true,
+    "ignorePatterns": [
+      "node_modules/**/*.js",
+      "dist/**/*.js"
+    ],
+    "parser": "@typescript-eslint/parser",
+    "plugins": [
+      "@typescript-eslint"
+    ],
+    "extends": [
+      "eslint:recommended",
+      "plugin:@typescript-eslint/recommended"
+    ],
+    "rules": {
+      "semi": [
+        "error",
+        "always"
+      ],
+      "object-curly-spacing": [
+        "error",
+        "always"
+      ],
+      "@typescript-eslint/consistent-type-definitions": [
+        "error",
+        "interface"
+      ],
+      "@typescript-eslint/array-type": [
+        "error",
+        {
+          "default": "array"
+        }
+      ],
+      "@typescript-eslint/consistent-type-assertions": "warn",
+      "@typescript-eslint/no-unused-vars": "warn",
+      "@typescript-eslint/naming-convention": [
+        "error",
+        {
+          "selector": [
+            "variable"
+          ],
+          "format": [
+            "camelCase",
+            "UPPER_CASE",
+            "PascalCase"
+          ]
+        }
+      ]
+    }
+  },
+  "devDependencies": {
+    "@types/connect-history-api-fallback": "^1.3.5",
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/jest": "^27.0.2",
+    "@types/node": "^16.11.1",
+    "@types/sqlite3": "^3.1.7",
+    "@types/supertest": "^2.0.11",
+    "@typescript-eslint/eslint-plugin": "^5.1.0",
+    "@typescript-eslint/parser": "^5.1.0",
+    "connect-history-api-fallback": "^1.6.0",
+    "eslint": "^8.0.1",
+    "jest": "^27.3.1",
+    "nodemon": "^2.0.19",
+    "supertest": "^6.1.6",
+    "ts-jest": "^27.0.7",
+    "ts-node": "^10.3.0",
+    "typescript": "^4.4.4"
+  },
+  "dependencies": {
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
+    "knex": "^0.95.11",
+    "sqlite3": "^5.0.9"
+  }
+}
