+++ src/api/tsconfig.json
@@ -6,7 +6,7 @@
     "jsx": "react", /* Specify JSX code generation: 'preserve', 'react-native', 'react', 'react-jsx' or 'react-jsxdev'. */
     "outDir": "./dist/api/", /* Redirect output structure to the directory. */
     "rootDirs": [
-      "./src/api"
+      "./"
     ],
     "strict": true, /* Enable all strict type-checking options. */
     "moduleResolution": "node", /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */
@@ -15,6 +15,6 @@
     "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
   },
   "include": [
-    "./src/api"
+    "./"
   ]
 }
