+++ src/client/src/util/constants.ts
@@ -1,0 +1,1 @@
+export const LOCAL_API_HOST = 'http://localhost:8080';
